+++ packages/headless/src/modules/media/rest/dto/uploadFile.dto.ts
